refactor(apply-inbound): split monolithic file into a module
The 1271-line `apply_inbound.rs` mixed three independent surfaces — the
inbound persistence/write ports, the blob materializer family, and the
top-level use case — alongside ~570 lines of tests. Reading the
materialize path (representation_index branch + free-standing file
branch + URL helpers) meant scrolling past the use case it lived next
to, and tests were impossible to navigate at this size.

Split into a directory module with one file per responsibility:

- `apply_inbound/mod.rs` — public types (`ApplyInboundInput`,
  `ApplyOutcome`, `ApplyInboundError`) and re-exports
- `apply_inbound/ports.rs` — `InboundCapture` / `InboundWrite` traits
  and their `CaptureClipboardUseCase` / `ClipboardWriteCoordinator`
  blanket impls
- `apply_inbound/materializer.rs` — `InboundBlobMaterializer` /
  `InboundBlobFetcher` traits + `FileCacheBlobMaterializer` plus the
  file-URL / sanitize / unique-name helpers it owns
- `apply_inbound/usecase.rs` — `ApplyInboundClipboardUseCase` and its
  `execute` orchestration
- `apply_inbound/tests.rs` — all mockall mocks + the 10 verdicts

External API unchanged: the `pub use` block in
`clipboard_sync/mod.rs` and the `lib.rs` root re-exports continue to
expose the same 8 symbols, so daemon / CLI / tauri callers do not
require any modification.

Pure refactor, no behavior change. All 10 apply_inbound tests pass and
`cargo check --workspace --tests` succeeds.
