+++ src-tauri/crates/uc-application/src/usecases/clipboard_sync/apply_inbound/tests.rs
@@ -1,0 +1,580 @@
+use std::sync::Arc;
+
+use anyhow::Result;
+use async_trait::async_trait;
+use bytes::Bytes;
+use mockall::predicate::*;
+
+use uc_core::ids::{DeviceId, EntryId, FormatId, RepresentationId};
+use uc_core::ports::blob::{BlobDigest, BlobTicket, PlaintextHash};
+use uc_core::ports::{ClipboardEntryRepositoryPort, PeerAddressError};
+use uc_core::{MimeType, ObservedClipboardRepresentation, SystemClipboardSnapshot};
+
+use crate::usecases::clipboard_sync::payload_codec::{
+    encode_snapshot_to_v3_bytes, encode_snapshot_with_blob_refs_to_v3_bytes, V3BlobRef,
+};
+
+use super::materializer::{FileCacheBlobMaterializer, InboundBlobFetcher, InboundBlobMaterializer};
+use super::ports::{InboundCapture, InboundWrite};
+use super::usecase::ApplyInboundClipboardUseCase;
+use super::{ApplyInboundError, ApplyInboundInput, ApplyOutcome};
+
+// ── mockall: the 3 collaborator surfaces ────────────────────────────
+
+mockall::mock! {
+    pub EntryRepo {}
+    #[async_trait]
+    impl ClipboardEntryRepositoryPort for EntryRepo {
+        async fn save_entry_and_selection(
+            &self,
+            entry: &uc_core::ClipboardEntry,
+            selection: &uc_core::ClipboardSelectionDecision,
+        ) -> Result<()>;
+        async fn get_entry(&self, entry_id: &EntryId) -> Result<Option<uc_core::ClipboardEntry>>;
+        async fn list_entries(&self, limit: usize, offset: usize) -> Result<Vec<uc_core::ClipboardEntry>>;
+        async fn touch_entry(&self, entry_id: &EntryId, active_time_ms: i64) -> Result<bool>;
+        async fn delete_entry(&self, entry_id: &EntryId) -> Result<()>;
+        async fn find_entry_id_by_snapshot_hash(&self, snapshot_hash: &str) -> Result<Option<EntryId>>;
+    }
+}
+
+mockall::mock! {
+    pub Capture {}
+    #[async_trait]
+    impl InboundCapture for Capture {
+        async fn capture(
+            &self,
+            preset_entry_id: EntryId,
+            snapshot: SystemClipboardSnapshot,
+        ) -> Result<Option<EntryId>>;
+    }
+}
+
+mockall::mock! {
+    pub Write {}
+    #[async_trait]
+    impl InboundWrite for Write {
+        async fn write(&self, snapshot: SystemClipboardSnapshot) -> Result<()>;
+    }
+}
+
+mockall::mock! {
+    pub BlobMaterializer {}
+    #[async_trait]
+    impl InboundBlobMaterializer for BlobMaterializer {
+        async fn materialize(
+            &self,
+            from_device: DeviceId,
+            receiver_entry_id: EntryId,
+            snapshot: SystemClipboardSnapshot,
+            blob_refs: Vec<V3BlobRef>,
+        ) -> Result<SystemClipboardSnapshot>;
+    }
+}
+
+mockall::mock! {
+    pub BlobFetcher {}
+    #[async_trait]
+    impl InboundBlobFetcher for BlobFetcher {
+        async fn fetch_blob(
+            &self,
+            command: crate::facade::blob_transfer::FetchBlobCommand,
+        ) -> Result<crate::facade::blob_transfer::FetchBlobResult>;
+    }
+}
+
+// ── helpers ─────────────────────────────────────────────────────────
+
+fn fixture_input(text: &str) -> (ApplyInboundInput, String) {
+    let snapshot = SystemClipboardSnapshot {
+        ts_ms: 1_700_000_000_000,
+        representations: vec![ObservedClipboardRepresentation::new(
+            RepresentationId::new(),
+            FormatId::from("text"),
+            Some(MimeType("text/plain".to_string())),
+            text.as_bytes().to_vec(),
+        )],
+    };
+    let (plaintext, content_hash) = encode_snapshot_to_v3_bytes(&snapshot).unwrap();
+    (
+        ApplyInboundInput {
+            from_device: DeviceId::new("peer-x"),
+            content_hash: content_hash.clone(),
+            plaintext,
+        },
+        content_hash,
+    )
+}
+
+fn build(
+    repo: MockEntryRepo,
+    capture: MockCapture,
+    write: MockWrite,
+) -> ApplyInboundClipboardUseCase {
+    ApplyInboundClipboardUseCase::new(Arc::new(repo), Arc::new(capture), Arc::new(write))
+}
+
+fn build_with_blob_materializer(
+    repo: MockEntryRepo,
+    capture: MockCapture,
+    write: MockWrite,
+    materializer: MockBlobMaterializer,
+) -> ApplyInboundClipboardUseCase {
+    ApplyInboundClipboardUseCase::new(Arc::new(repo), Arc::new(capture), Arc::new(write))
+        .with_blob_materializer(Arc::new(materializer))
+}
+
+// ── verdicts ────────────────────────────────────────────────────────
+
+/// Verdict 1 — happy path: dedup miss → decode → capture → write →
+/// `Applied { entry_id }`. mockall asserts: dedup query once with
+/// the input hash, capture once, write once.
+#[tokio::test]
+async fn applied_on_new_content() {
+    let (input, hash) = fixture_input("hello phase3");
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .with(eq(hash.clone()))
+        .times(1)
+        .returning(|_| Ok(None));
+
+    let mut capture = MockCapture::new();
+    capture
+        .expect_capture()
+        .times(1)
+        .returning(|_, _| Ok(Some(EntryId::from("entry-new"))));
+
+    let mut write = MockWrite::new();
+    write.expect_write().times(1).returning(|_| Ok(()));
+
+    let uc = build(repo, capture, write);
+    let outcome = uc.execute(input).await.expect("happy path returns ok");
+    assert_eq!(
+        outcome,
+        ApplyOutcome::Applied {
+            entry_id: EntryId::from("entry-new")
+        }
+    );
+}
+
+/// Verdict 2 — dedup hit: returns `DuplicateSkipped` and **does
+/// not** call capture or write. Critical correctness property —
+/// repeated dispatches from a peer must not double-write the user's
+/// OS clipboard (Phase 3 acceptance #4).
+#[tokio::test]
+async fn duplicate_skipped_when_hash_already_local() {
+    let (input, hash) = fixture_input("already-here");
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .with(eq(hash.clone()))
+        .times(1)
+        .returning(|_| Ok(Some(EntryId::from("entry-existing"))));
+
+    // Zero expectations on capture + write — mockall panics on Drop
+    // if either gets called. This pins the no-side-effect contract.
+    let capture = MockCapture::new();
+    let write = MockWrite::new();
+
+    let uc = build(repo, capture, write);
+    let outcome = uc.execute(input).await.expect("dedup path ok");
+    assert_eq!(
+        outcome,
+        ApplyOutcome::DuplicateSkipped {
+            content_hash: hash,
+            existing_entry_id: EntryId::from("entry-existing"),
+        }
+    );
+}
+
+/// Verdict 3 — corrupt envelope returns `DecodeFailed`, no panic, no
+/// capture, no write. Daemon's ingest loop keeps running.
+#[tokio::test]
+async fn decode_failed_on_truncated_envelope() {
+    let input = ApplyInboundInput {
+        from_device: DeviceId::new("peer-broken"),
+        content_hash: "blake3v1:00".to_string(),
+        plaintext: Bytes::from_static(b"not a valid V3 envelope"),
+    };
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .times(1)
+        .returning(|_| Ok(None));
+    let capture = MockCapture::new();
+    let write = MockWrite::new();
+
+    let uc = build(repo, capture, write);
+    let outcome = uc.execute(input).await.expect("DecodeFailed is Ok variant");
+    match outcome {
+        ApplyOutcome::DecodeFailed { reason } => {
+            assert!(
+                reason.contains("decode V3 envelope"),
+                "reason should mention V3 decode, got: {reason}"
+            );
+        }
+        other => panic!("expected DecodeFailed, got {other:?}"),
+    }
+}
+
+/// Verdict 4 — capture returns Ok(None) (shouldn't happen for
+/// RemotePush but guard it anyway): mapped to
+/// `ApplyInboundError::Internal`. Write must NOT fire.
+#[tokio::test]
+async fn capture_returning_none_maps_to_internal_error() {
+    let (input, _) = fixture_input("orphan");
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .times(1)
+        .returning(|_| Ok(None));
+
+    let mut capture = MockCapture::new();
+    capture.expect_capture().times(1).returning(|_, _| Ok(None));
+
+    // Zero expectations on write.
+    let write = MockWrite::new();
+
+    let uc = build(repo, capture, write);
+    let err = uc
+        .execute(input)
+        .await
+        .expect_err("Ok(None) from capture must surface as error");
+    match err {
+        ApplyInboundError::Internal(msg) => {
+            assert!(
+                msg.contains("RemotePush"),
+                "internal message should reference origin, got: {msg}"
+            );
+        }
+        other => panic!("expected Internal, got {other:?}"),
+    }
+}
+
+/// Verdict 5 — write coordinator failure surfaces as
+/// `WriteCoordinator` error. Capture has already committed (the
+/// entry stays in DB; manual cleanup is the daemon operator's job).
+/// Pin this trade-off so a future refactor doesn't silently start
+/// rolling back persistence on write failure.
+#[tokio::test]
+async fn write_failure_surfaces_after_capture_commits() {
+    let (input, _) = fixture_input("write-will-fail");
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .times(1)
+        .returning(|_| Ok(None));
+
+    let mut capture = MockCapture::new();
+    capture
+        .expect_capture()
+        .times(1)
+        .returning(|_, _| Ok(Some(EntryId::from("entry-committed"))));
+
+    let mut write = MockWrite::new();
+    write
+        .expect_write()
+        .times(1)
+        .returning(|_| Err(anyhow::anyhow!("OS clipboard locked")));
+
+    let uc = build(repo, capture, write);
+    let err = uc
+        .execute(input)
+        .await
+        .expect_err("write failure must surface");
+    match err {
+        ApplyInboundError::WriteCoordinator(msg) => {
+            assert!(
+                msg.contains("OS clipboard locked"),
+                "underlying error should propagate, got: {msg}"
+            );
+        }
+        other => panic!("expected WriteCoordinator, got {other:?}"),
+    }
+}
+
+/// Verdict 6 — dedup query failure surfaces as `DedupQuery`. No
+/// decode, no capture, no write — failing closed is the conservative
+/// choice (we'd rather lose an inbound frame than risk a corrupt
+/// double-write under unknown DB state).
+#[tokio::test]
+async fn dedup_query_failure_short_circuits() {
+    let (input, _) = fixture_input("dedup-broken");
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .times(1)
+        .returning(|_| {
+            Err(anyhow::Error::from(PeerAddressError::Internal(
+                "db down".to_string(),
+            )))
+        });
+    let capture = MockCapture::new();
+    let write = MockWrite::new();
+
+    let uc = build(repo, capture, write);
+    let err = uc.execute(input).await.expect_err("dedup error propagates");
+    match err {
+        ApplyInboundError::DedupQuery(_) => {}
+        other => panic!("expected DedupQuery, got {other:?}"),
+    }
+}
+
+/// Verdict 7 — 入站 blob refs 会先本地化,再进入 capture 和剪贴板写入。
+/// capture/write mock 校验收到的是改写后的本机 file URI,不是发送端原始路径。
+#[tokio::test]
+async fn materializes_blob_refs_before_capture_and_write() {
+    let original = SystemClipboardSnapshot {
+        ts_ms: 1_700_000_000_000,
+        representations: vec![ObservedClipboardRepresentation::new(
+            RepresentationId::new(),
+            FormatId::from("files"),
+            Some(MimeType("text/uri-list".to_string())),
+            b"file:///sender/original.txt\n".to_vec(),
+        )],
+    };
+    let blob_ref = V3BlobRef {
+        ticket: BlobTicket::from_bytes(vec![9, 8, 7]),
+        entry_id: EntryId::from("entry-remote"),
+        filename: Some("original.txt".to_string()),
+        mime: Some("text/plain".to_string()),
+        size_bytes: 13,
+        representation_index: None,
+    };
+    let (plaintext, content_hash) =
+        encode_snapshot_with_blob_refs_to_v3_bytes(&original, &[blob_ref.clone()]).unwrap();
+    let input = ApplyInboundInput {
+        from_device: DeviceId::new("peer-x"),
+        content_hash: content_hash.clone(),
+        plaintext,
+    };
+
+    let mut repo = MockEntryRepo::new();
+    repo.expect_find_entry_id_by_snapshot_hash()
+        .with(eq(content_hash))
+        .times(1)
+        .returning(|_| Ok(None));
+
+    let mut materializer = MockBlobMaterializer::new();
+    materializer
+        .expect_materialize()
+        .times(1)
+        .withf(move |_from_device, _receiver_entry_id, snapshot, refs| {
+            snapshot.representations[0].bytes == b"file:///sender/original.txt\n"
+                && refs == &vec![blob_ref.clone()]
+        })
+        .returning(|_from_device, _receiver_entry_id, mut snapshot, _| {
+            snapshot.representations[0].bytes = b"file:///local/cache/original.txt\n".to_vec();
+            Ok(snapshot)
+        });
+
+    let assert_local_file = |snapshot: &SystemClipboardSnapshot| {
+        snapshot.representations[0].bytes == b"file:///local/cache/original.txt\n"
+    };
+    let mut capture = MockCapture::new();
+    capture
+        .expect_capture()
+        .withf(move |_preset_entry_id, snapshot| assert_local_file(snapshot))
+        .times(1)
+        .returning(|_, _| Ok(Some(EntryId::from("entry-new"))));
+
+    let mut write = MockWrite::new();
+    write
+        .expect_write()
+        .withf(move |snapshot| assert_local_file(snapshot))
+        .times(1)
+        .returning(|_| Ok(()));
+
+    let uc = build_with_blob_materializer(repo, capture, write, materializer);
+    let outcome = uc.execute(input).await.expect("blob materialize path ok");
+    assert_eq!(
+        outcome,
+        ApplyOutcome::Applied {
+            entry_id: EntryId::from("entry-new")
+        }
+    );
+}
+
+/// Verdict 8 — 真实文件缓存 materializer 会拉取 blob 内容,写入接收端缓存目录,
+/// 并把 file-list 表示改写为本机 `file://` URI。
+#[tokio::test]
+async fn file_cache_blob_materializer_writes_file_and_rewrites_file_uri_list() {
+    let cache_dir = tempfile::tempdir().expect("tempdir");
+    let entry_id = EntryId::from("entry-file");
+    let ticket = BlobTicket::from_bytes(vec![1, 2, 3]);
+    let blob_ref = V3BlobRef {
+        ticket: ticket.clone(),
+        entry_id: entry_id.clone(),
+        filename: Some("report.txt".to_string()),
+        mime: Some("text/plain".to_string()),
+        size_bytes: 11,
+        representation_index: None,
+    };
+    let snapshot = SystemClipboardSnapshot {
+        ts_ms: 1,
+        representations: vec![ObservedClipboardRepresentation::new(
+            RepresentationId::new(),
+            FormatId::from("files"),
+            Some(MimeType("text/uri-list".to_string())),
+            b"file:///sender/report.txt\n".to_vec(),
+        )],
+    };
+
+    let mut fetcher = MockBlobFetcher::new();
+    fetcher
+        .expect_fetch_blob()
+        .times(1)
+        .withf(move |command| command.entry_id == entry_id && command.ticket == ticket)
+        .returning(|command| {
+            Ok(crate::facade::blob_transfer::FetchBlobResult {
+                plaintext: Bytes::from_static(b"hello world"),
+                entry_id: command.entry_id,
+                plaintext_hash: PlaintextHash::from_bytes([0; 32]),
+                digest: BlobDigest::from_bytes([1; 32]),
+            })
+        });
+
+    let materializer =
+        FileCacheBlobMaterializer::new(Arc::new(fetcher), cache_dir.path().to_path_buf());
+    let rewritten = materializer
+        .materialize(
+            DeviceId::new("peer-x"),
+            EntryId::from("entry-receiver"),
+            snapshot,
+            vec![blob_ref],
+        )
+        .await
+        .expect("materialize should succeed");
+
+    let uri_list = String::from_utf8(rewritten.representations[0].bytes.clone())
+        .expect("uri-list should be UTF-8");
+    assert!(uri_list.starts_with("file://"));
+    assert!(uri_list.ends_with("/report.txt\n"));
+    assert!(!uri_list.contains("/sender/"));
+
+    let local_url = url::Url::parse(uri_list.trim()).expect("valid file URL");
+    let local_path = local_url.to_file_path().expect("file URL to path");
+    let bytes = tokio::fs::read(local_path)
+        .await
+        .expect("materialized file should exist");
+    assert_eq!(bytes, b"hello world");
+}
+
+/// Verdict 9 —— representation_index 路径：blob ref 携带索引时,materializer
+/// 把 fetched bytes 灌回 envelope 主体里对应索引的 rep,而不是写到 cache_dir
+/// 当 file 处理。这条路径是 oversized image 跨设备同步的关键。
+#[tokio::test]
+async fn file_cache_blob_materializer_inlines_representation_bound_blob_into_rep() {
+    let cache_dir = tempfile::tempdir().expect("tempdir");
+    let entry_id = EntryId::from("entry-img");
+    let ticket = BlobTicket::from_bytes(vec![7, 7, 7]);
+    let blob_ref = V3BlobRef {
+        ticket: ticket.clone(),
+        entry_id: entry_id.clone(),
+        filename: None,
+        mime: Some("image/png".to_string()),
+        size_bytes: 5,
+        representation_index: Some(0),
+    };
+    // Sender drained `bytes` to empty when publishing — receiver decode
+    // mirrors that empty-rep state until materialization runs.
+    let snapshot = SystemClipboardSnapshot {
+        ts_ms: 1,
+        representations: vec![ObservedClipboardRepresentation::new(
+            RepresentationId::new(),
+            FormatId::from("image"),
+            Some(MimeType("image/png".to_string())),
+            Vec::new(),
+        )],
+    };
+
+    let mut fetcher = MockBlobFetcher::new();
+    fetcher
+        .expect_fetch_blob()
+        .times(1)
+        .withf(move |command| command.entry_id == entry_id && command.ticket == ticket)
+        .returning(|command| {
+            Ok(crate::facade::blob_transfer::FetchBlobResult {
+                plaintext: Bytes::from_static(b"\x89PNG\x0d"),
+                entry_id: command.entry_id,
+                plaintext_hash: PlaintextHash::from_bytes([0; 32]),
+                digest: BlobDigest::from_bytes([1; 32]),
+            })
+        });
+
+    let materializer =
+        FileCacheBlobMaterializer::new(Arc::new(fetcher), cache_dir.path().to_path_buf());
+    let materialized = materializer
+        .materialize(
+            DeviceId::new("peer-x"),
+            EntryId::from("entry-receiver"),
+            snapshot,
+            vec![blob_ref],
+        )
+        .await
+        .expect("representation-bound materialize should succeed");
+
+    assert_eq!(materialized.representations.len(), 1);
+    assert_eq!(materialized.representations[0].bytes, b"\x89PNG\x0d");
+    assert_eq!(materialized.representations[0].format_id.as_ref(), "image");
+
+    let mut entries = tokio::fs::read_dir(cache_dir.path())
+        .await
+        .expect("read cache_dir");
+    assert!(
+        entries.next_entry().await.expect("read entry").is_none(),
+        "cache_dir must be empty for representation-bound refs"
+    );
+}
+
+/// Verdict 10 —— representation_index 越界时,materializer 必须显式报错而不是
+/// silently 落到 file 路径或 panic。这个 guard 防止协议不一致的对端把消息
+/// 灌进错误的 rep slot。
+#[tokio::test]
+async fn file_cache_blob_materializer_rejects_out_of_bounds_representation_index() {
+    let cache_dir = tempfile::tempdir().expect("tempdir");
+    let blob_ref = V3BlobRef {
+        ticket: BlobTicket::from_bytes(vec![1]),
+        entry_id: EntryId::from("entry-bad"),
+        filename: None,
+        mime: Some("image/png".to_string()),
+        size_bytes: 1,
+        representation_index: Some(5),
+    };
+    let snapshot = SystemClipboardSnapshot {
+        ts_ms: 1,
+        representations: vec![ObservedClipboardRepresentation::new(
+            RepresentationId::new(),
+            FormatId::from("image"),
+            Some(MimeType("image/png".to_string())),
+            Vec::new(),
+        )],
+    };
+
+    let mut fetcher = MockBlobFetcher::new();
+    fetcher.expect_fetch_blob().times(1).returning(|command| {
+        Ok(crate::facade::blob_transfer::FetchBlobResult {
+            plaintext: Bytes::from_static(b"x"),
+            entry_id: command.entry_id,
+            plaintext_hash: PlaintextHash::from_bytes([0; 32]),
+            digest: BlobDigest::from_bytes([1; 32]),
+        })
+    });
+
+    let materializer =
+        FileCacheBlobMaterializer::new(Arc::new(fetcher), cache_dir.path().to_path_buf());
+    let err = materializer
+        .materialize(
+            DeviceId::new("peer-x"),
+            EntryId::from("entry-receiver"),
+            snapshot,
+            vec![blob_ref],
+        )
+        .await
+        .expect_err("out-of-bounds index should fail");
+    assert!(
+        err.to_string().contains("out of bounds"),
+        "error must mention out-of-bounds context: {err}"
+    );
+}
